+++ packages/delivery-engine/src/index.js
@@ -1,3 +1,4 @@
 export * from './geo.js';
 export * from './serviceability.js';
 export * from './dispatch.js';
+export * from './eta.js';